Split out type info for general slang types.
