+++ source/slang/slang.vcxproj
@@ -289,6 +289,7 @@
     <ClInclude Include="slang-serialize-ir-types.h" />
     <ClInclude Include="slang-serialize-ir.h" />
     <ClInclude Include="slang-serialize-reflection.h" />
+    <ClInclude Include="slang-serialize-type-info.h" />
     <ClInclude Include="slang-serialize-types.h" />
     <ClInclude Include="slang-serialize.h" />
     <ClInclude Include="slang-source-loc.h" />
